+++ src/main/java/com/ghgande/j2mod/modbus/facade/ModbusSerialMaster.java
@@ -91,7 +91,7 @@ public class ModbusSerialMaster extends AbstractModbusMaster {
      *
      * @throws Exception if the connection cannot be established.
      */
-    public void connect() throws Exception {
+    public synchronized void connect() throws Exception {
         if (connection != null && !connection.isOpen()) {
             connection.open();
             transaction = connection.getModbusTransport().createTransaction();
@@ -103,7 +103,7 @@ public class ModbusSerialMaster extends AbstractModbusMaster {
     /**
      * Disconnects this <tt>ModbusSerialMaster</tt> from the slave.
      */
-    public void disconnect() {
+    public synchronized void disconnect() {
         if (connection != null && connection.isOpen()) {
             connection.close();
             transaction = null;
@@ -112,7 +112,7 @@ public class ModbusSerialMaster extends AbstractModbusMaster {
     }
 
     @Override
-    public void setTimeout(int timeout) {
+    public synchronized void setTimeout(int timeout) {
         super.setTimeout(timeout);
         if (connection != null) {
             connection.setTimeout(timeout);