+++ src/main/java/com/ghgande/j2mod/modbus/facade/ModbusUDPMaster.java
@@ -87,7 +87,7 @@ public class ModbusUDPMaster extends AbstractModbusMaster {
      *
      * @throws Exception if the connection cannot be established.
      */
-    public void connect() throws Exception {
+    public synchronized void connect() throws Exception {
         if (connection != null && !connection.isConnected()) {
             connection.connect();
             transaction = connection.getModbusTransport().createTransaction();
@@ -98,7 +98,7 @@ public class ModbusUDPMaster extends AbstractModbusMaster {
     /**
      * Disconnects this <tt>ModbusTCPMaster</tt> from the slave.
      */
-    public void disconnect() {
+    public synchronized void disconnect() {
         if (connection != null && connection.isConnected()) {
             connection.close();
             transaction = null;
@@ -107,7 +107,7 @@ public class ModbusUDPMaster extends AbstractModbusMaster {
     }
 
     @Override
-    public void setTimeout(int timeout) {
+    public synchronized void setTimeout(int timeout) {
         super.setTimeout(timeout);
         if (connection != null) {
             connection.setTimeout(timeout);