+++ RELEASE_NOTES.md
@@ -182,3 +182,4 @@ _(**NOT BACKWARDS COMPATIBLE**)_
 
 ## Version 2.5.8
 * Upgraded jserialcomm to v2.5.3
+* ClassCastException when timeout?? #96 - made all Master facade methods thread safe